designed login and register form, added need of authorization to post products

+++ backend/src/main/java/com/springBootProjects/SimpleWebApplication/service/ProductService.java
@@ -1,7 +1,10 @@
 package com.springBootProjects.SimpleWebApplication.service;
 
+import com.springBootProjects.SimpleWebApplication.exception.UserNotLoggedInException;
 import com.springBootProjects.SimpleWebApplication.model.Product;
+import com.springBootProjects.SimpleWebApplication.model.User;
 import com.springBootProjects.SimpleWebApplication.repository.ProductRepository;
+import jakarta.persistence.EntityNotFoundException;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -9,6 +12,7 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.multipart.MultipartFile;
 
 import java.io.IOException;
+import java.nio.file.AccessDeniedException;
 import java.util.List;
 
 @Service
@@ -17,33 +21,46 @@ public class ProductService {
     @Autowired
     private ProductRepository productRepository;
 
+    @Autowired
+    private UserService userService;
+
     public List<Product> getProducts(){
         return productRepository.findAll();
     }
 
-    public Product getProductById(int id){
-        return productRepository.findById(id).orElse(null);
+    public Product getProductById(int id) {
+        return productRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("Produkt wurde nicht gefunden"));
     }
 
 
-    public Product addProduct(Product product) throws IOException {
+    public Product addProduct(Product product) {
+        User currentUser = userService.getAuthenticatedUser();
+        product.setUser(currentUser);
         return productRepository.save(product);
     }
 
-    public Product updateProduct(int id, Product product) throws IOException {
-        // 1. Check whether id is correct
-        // 2. Check whether an entry with the provided id exists
-        // If not, don't insert it.
-        if (product.getId() != id || !productRepository.existsById(id)) {
-            return null;
+
+    public Product updateProduct(int id, Product product) {
+        Product existingProduct = getProductById(id);
+
+        // check if user posted the product in the first place
+        User currentUser = userService.getAuthenticatedUser();
+
+        if(existingProduct.getUser().getId() != currentUser.getId()){
+            throw new UserNotLoggedInException("Sie sind nicht der Besitzer dieses Produkts");
         }
         return productRepository.save(product);
     }
 
     public Product deleteProduct(int id) {
-        Product p = this.getProductById(id);
+        Product existingProduct = getProductById(id);
+        // check if user posted the product in the first place
+        User currentUser = userService.getAuthenticatedUser();
+        if(existingProduct.getUser().getId() != currentUser.getId()){
+            throw new UserNotLoggedInException("Sie sind nicht der Besitzer dieses Produkts");
+        }
         productRepository.deleteById(id);
-        return p;
+        return existingProduct;
     }
 
 }
